Split off network and api logic into modules
