+++ testkit/src/lib.rs
@@ -136,224 +136,33 @@ extern crate log;
 use futures::Stream;
 use futures::executor::{self, Spawn};
 use futures::sync::mpsc;
-use iron::{IronError, Handler, Chain};
-use iron::headers::{ContentType, Headers};
-use iron::status::StatusClass;
-use iron_test::{request, response};
-use serde::{Deserialize, Serialize};
 
 use std::collections::BTreeMap;
 use std::sync::{Arc, RwLock, RwLockReadGuard};
 use std::fmt;
 
-use exonum::blockchain::{Blockchain, ConsensusConfig, GenesisConfig, Schema as CoreSchema,
-                         Service, SharedNodeState, StoredConfiguration, Transaction, ValidatorKeys};
-use exonum::crypto::{self, CryptoHash};
-use exonum::helpers::{Height, Round, ValidatorId};
-use exonum::messages::{Precommit, Propose};
-use exonum::node::{ApiSender, ExternalMessage, State as NodeState, TransactionSend, TxPool,
-                   NodeApiConfig, create_public_api_handler, create_private_api_handler};
+use exonum::blockchain::{Blockchain, Schema as CoreSchema, Service, StoredConfiguration,
+                         Transaction};
+use exonum::crypto;
+use exonum::helpers::{Height, ValidatorId};
+use exonum::node::{ApiSender, ExternalMessage, State as NodeState, TxPool, NodeApiConfig};
 use exonum::storage::{MemoryDB, Snapshot};
 
 #[macro_use]
 mod macros;
+mod api;
 mod checkpoint_db;
 pub mod compare;
 mod greedy_fold;
+mod network;
 
 #[doc(hidden)]
 pub use greedy_fold::GreedilyFoldable;
+pub use api::{ApiKind, TestKitApi};
 pub use compare::ComparableSnapshot;
+pub use network::{TestNetwork, TestNode, TestNetworkConfiguration};
 
 use checkpoint_db::{CheckpointDb, CheckpointDbHandler};
-
-/// Emulated test network.
-#[derive(Debug)]
-pub struct TestNetwork {
-    us: TestNode,
-    validators: Vec<TestNode>,
-}
-
-impl TestNetwork {
-    /// Creates a new emulated network.
-    pub fn new(validator_count: u16) -> Self {
-        let validators = (0..validator_count)
-            .map(ValidatorId)
-            .map(TestNode::new_validator)
-            .collect::<Vec<_>>();
-
-        let us = validators[0].clone();
-        TestNetwork { validators, us }
-    }
-
-    /// Returns the node in the emulated network, from whose perspective the testkit operates.
-    pub fn us(&self) -> &TestNode {
-        &self.us
-    }
-
-    /// Returns a slice of all validators in the network.
-    pub fn validators(&self) -> &[TestNode] {
-        &self.validators
-    }
-
-    /// Returns config encoding the network structure usable for creating the genesis block of
-    /// a blockchain.
-    pub fn genesis_config(&self) -> GenesisConfig {
-        GenesisConfig::new(self.validators.iter().map(TestNode::public_keys))
-    }
-
-    /// Updates the test network by the new set of nodes.
-    pub fn update<I: IntoIterator<Item = TestNode>>(&mut self, mut us: TestNode, validators: I) {
-        let validators = validators
-            .into_iter()
-            .enumerate()
-            .map(|(id, mut validator)| {
-                let validator_id = ValidatorId(id as u16);
-                validator.change_role(Some(validator_id));
-                if us.public_keys().consensus_key == validator.public_keys().consensus_key {
-                    us.change_role(Some(validator_id));
-                }
-                validator
-            })
-            .collect::<Vec<_>>();
-        self.validators = validators;
-        self.us.clone_from(&us);
-    }
-
-    /// Returns service public key of the validator with given id.
-    pub fn service_public_key_of(&self, id: ValidatorId) -> Option<&crypto::PublicKey> {
-        self.validators().get(id.0 as usize).map(|x| {
-            &x.service_public_key
-        })
-    }
-
-    /// Returns consensus public key of the validator with given id.
-    pub fn consensus_public_key_of(&self, id: ValidatorId) -> Option<&crypto::PublicKey> {
-        self.validators().get(id.0 as usize).map(|x| {
-            &x.consensus_public_key
-        })
-    }
-}
-
-/// An emulated node in the test network.
-#[derive(Debug, Clone, PartialEq)]
-pub struct TestNode {
-    consensus_secret_key: crypto::SecretKey,
-    consensus_public_key: crypto::PublicKey,
-    service_secret_key: crypto::SecretKey,
-    service_public_key: crypto::PublicKey,
-    validator_id: Option<ValidatorId>,
-}
-
-impl TestNode {
-    /// Creates a new auditor.
-    pub fn new_auditor() -> Self {
-        let (consensus_public_key, consensus_secret_key) = crypto::gen_keypair();
-        let (service_public_key, service_secret_key) = crypto::gen_keypair();
-
-        TestNode {
-            consensus_secret_key,
-            consensus_public_key,
-            service_secret_key,
-            service_public_key,
-            validator_id: None,
-        }
-    }
-
-    /// Creates a new validator with the given id.
-    pub fn new_validator(validator_id: ValidatorId) -> Self {
-        let (consensus_public_key, consensus_secret_key) = crypto::gen_keypair();
-        let (service_public_key, service_secret_key) = crypto::gen_keypair();
-
-        TestNode {
-            consensus_secret_key,
-            consensus_public_key,
-            service_secret_key,
-            service_public_key,
-            validator_id: Some(validator_id),
-        }
-    }
-
-    /// Constructs a new node from the given keypairs.
-    pub fn from_parts(
-        consensus_keypair: (crypto::PublicKey, crypto::SecretKey),
-        service_keypair: (crypto::PublicKey, crypto::SecretKey),
-        validator_id: Option<ValidatorId>,
-    ) -> TestNode {
-        TestNode {
-            consensus_public_key: consensus_keypair.0,
-            consensus_secret_key: consensus_keypair.1,
-            service_public_key: service_keypair.0,
-            service_secret_key: service_keypair.1,
-            validator_id,
-        }
-    }
-
-    /// Creates a `Propose` message signed by this validator.
-    pub fn create_propose(
-        &self,
-        height: Height,
-        last_hash: &crypto::Hash,
-        tx_hashes: &[crypto::Hash],
-    ) -> Propose {
-        Propose::new(
-            self.validator_id.expect(
-                "An attempt to create propose from a non-validator node.",
-            ),
-            height,
-            Round::first(),
-            last_hash,
-            tx_hashes,
-            &self.consensus_secret_key,
-        )
-    }
-
-    /// Creates a `Precommit` message signed by this validator.
-    pub fn create_precommit(&self, propose: &Propose, block_hash: &crypto::Hash) -> Precommit {
-        use std::time::SystemTime;
-
-        Precommit::new(
-            self.validator_id.expect(
-                "An attempt to create propose from a non-validator node.",
-            ),
-            propose.height(),
-            propose.round(),
-            &propose.hash(),
-            block_hash,
-            SystemTime::now(),
-            &self.consensus_secret_key,
-        )
-    }
-
-    /// Returns public keys of the node.
-    pub fn public_keys(&self) -> ValidatorKeys {
-        ValidatorKeys {
-            consensus_key: self.consensus_public_key,
-            service_key: self.service_public_key,
-        }
-    }
-
-    /// Returns the current validator id of node if it is validator of the test network.
-    pub fn validator_id(&self) -> Option<ValidatorId> {
-        self.validator_id
-    }
-
-    /// Changes node role.
-    pub fn change_role(&mut self, role: Option<ValidatorId>) {
-        self.validator_id = role;
-    }
-
-    /// Returns the service keypair.
-    pub fn service_keypair(&self) -> (&crypto::PublicKey, &crypto::SecretKey) {
-        (&self.service_public_key, &self.service_secret_key)
-    }
-}
-
-impl From<TestNode> for ValidatorKeys {
-    fn from(node: TestNode) -> Self {
-        node.public_keys()
-    }
-}
 
 /// Builder for `TestKit`.
 ///
@@ -391,8 +200,8 @@ impl From<TestNode> for ValidatorKeys {
 /// # }
 /// ```
 pub struct TestKitBuilder {
-    us: TestNode,
-    validators: Vec<TestNode>,
+    our_validator_id: Option<ValidatorId>,
+    validator_count: Option<u16>,
     services: Vec<Box<Service>>,
     logger: bool,
 }
@@ -400,8 +209,13 @@ pub struct TestKitBuilder {
 impl fmt::Debug for TestKitBuilder {
     fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
         f.debug_struct("TestKitBuilder")
-            .field("us", &self.us)
-            .field("validators", &self.validators)
+            .field(
+                "us",
+                &self.our_validator_id.map_or("Auditor".to_string(), |id| {
+                    format!("Validator #{}", id.0)
+                }),
+            )
+            .field("validator_count", &self.validator_count)
             .field(
                 "services",
                 &self.services
@@ -409,6 +223,7 @@ impl fmt::Debug for TestKitBuilder {
                     .map(|x| x.service_name())
                     .collect::<Vec<_>>(),
             )
+            .field("logger", &self.logger)
             .finish()
     }
 }
@@ -416,36 +231,31 @@ impl fmt::Debug for TestKitBuilder {
 impl TestKitBuilder {
     /// Creates testkit for the validator node.
     pub fn validator() -> Self {
-        let us = TestNode::new_validator(ValidatorId(0));
         TestKitBuilder {
-            validators: vec![us.clone()],
+            validator_count: None,
+            our_validator_id: Some(ValidatorId(0)),
             services: Vec::new(),
-            us,
             logger: false,
         }
     }
 
     /// Creates testkit for the auditor node.
     pub fn auditor() -> Self {
-        let us = TestNode::new_auditor();
         TestKitBuilder {
-            validators: vec![TestNode::new_validator(ValidatorId(0))],
+            validator_count: None,
+            our_validator_id: None,
             services: Vec::new(),
-            us,
             logger: false,
         }
     }
 
     /// Sets the number of validator nodes in the test network.
-    pub fn with_validators(mut self, validators_count: u16) -> Self {
+    pub fn with_validators(mut self, validator_count: u16) -> Self {
         assert!(
-            validators_count > 0,
-            "At least one validator should be present in the network."
+            self.validator_count.is_none(),
+            "Number of validators is already specified"
         );
-        let additional_validators = (self.validators.len() as u16..validators_count)
-            .map(ValidatorId)
-            .map(TestNode::new_validator);
-        self.validators.extend(additional_validators);
+        self.validator_count = Some(validator_count);
         self
     }
 
@@ -472,10 +282,7 @@ impl TestKitBuilder {
         crypto::init();
         TestKit::assemble(
             self.services,
-            TestNetwork {
-                us: self.us,
-                validators: self.validators,
-            },
+            TestNetwork::with_our_role(self.our_validator_id, self.validator_count.unwrap_or(1)),
         )
     }
 }
@@ -757,10 +564,7 @@ impl TestKit {
                 Committed(cfg_proposal) => {
                     if cfg_proposal.actual_from() == actual_from {
                         // Modify the self configuration
-                        self.network_mut().update(
-                            cfg_proposal.us,
-                            cfg_proposal.validators,
-                        );
+                        self.network_mut().update_configuration(cfg_proposal);
                     } else {
                         self.cfg_proposal = Some(Committed(cfg_proposal));
                     }
@@ -882,7 +686,7 @@ impl TestKit {
     ///
     /// - Panics if validator with the given id is absent in test network.
     pub fn validator(&self, id: ValidatorId) -> &TestNode {
-        &self.network.validators[id.0 as usize]
+        &self.network.validators()[id.0 as usize]
     }
 
     /// Returns sufficient number of validators for the Byzantine Fault Tolerance consensus.
@@ -899,7 +703,7 @@ impl TestKit {
 
     /// Returns the leader on the current height. At the moment first validator.
     pub fn leader(&self) -> &TestNode {
-        &self.network().validators[0]
+        &self.network().validators()[0]
     }
 
     /// Returns the reference to test network.
@@ -917,11 +721,7 @@ impl TestKit {
     /// `commit_configuration_change` method.
     pub fn configuration_change_proposal(&self) -> TestNetworkConfiguration {
         let stored_configuration = CoreSchema::new(&self.snapshot()).actual_configuration();
-        TestNetworkConfiguration::from_parts(
-            self.network().us().clone(),
-            self.network().validators().into(),
-            stored_configuration,
-        )
+        TestNetworkConfiguration::new(self.network(), stored_configuration)
     }
 
     /// Adds a new configuration proposal. Remember, to add this proposal to the blockchain,
@@ -982,6 +782,7 @@ impl TestKit {
     /// ```
     pub fn commit_configuration_change(&mut self, proposal: TestNetworkConfiguration) {
         use self::ConfigurationProposalState::*;
+
         assert!(
             self.height() < proposal.actual_from(),
             "The `actual_from` height should be greater than the current."
@@ -999,385 +800,11 @@ impl TestKit {
     }
 }
 
-/// A configuration of the test network.
-#[derive(Debug)]
-pub struct TestNetworkConfiguration {
-    us: TestNode,
-    validators: Vec<TestNode>,
-    stored_configuration: StoredConfiguration,
-}
-
 // A new configuration proposal state.
 #[derive(Debug)]
 enum ConfigurationProposalState {
     Uncommitted(TestNetworkConfiguration),
     Committed(TestNetworkConfiguration),
-}
-
-impl TestNetworkConfiguration {
-    fn from_parts(
-        us: TestNode,
-        validators: Vec<TestNode>,
-        mut stored_configuration: StoredConfiguration,
-    ) -> Self {
-        let prev_hash = exonum::crypto::CryptoHash::hash(&stored_configuration);
-        stored_configuration.previous_cfg_hash = prev_hash;
-        TestNetworkConfiguration {
-            us,
-            validators,
-            stored_configuration,
-        }
-    }
-
-    /// Returns the node from whose perspective the testkit operates.
-    pub fn us(&self) -> &TestNode {
-        &self.us
-    }
-
-    /// Modifies the node from whose perspective the testkit operates.
-    pub fn set_us(&mut self, us: TestNode) {
-        self.us = us;
-        self.update_our_role();
-    }
-
-    /// Returns the test network validators.
-    pub fn validators(&self) -> &[TestNode] {
-        self.validators.as_ref()
-    }
-
-    /// Returns the current consensus configuration.
-    pub fn consensus_configuration(&self) -> &ConsensusConfig {
-        &self.stored_configuration.consensus
-    }
-
-    /// Return the height, starting from which this configuration becomes actual.
-    pub fn actual_from(&self) -> Height {
-        self.stored_configuration.actual_from
-    }
-
-    /// Modifies the height, starting from which this configuration becomes actual.
-    pub fn set_actual_from(&mut self, actual_from: Height) {
-        self.stored_configuration.actual_from = actual_from;
-    }
-
-    /// Modifies the current consensus configuration.
-    pub fn set_consensus_configuration(&mut self, consensus: ConsensusConfig) {
-        self.stored_configuration.consensus = consensus;
-    }
-
-    /// Modifies the validators list.
-    pub fn set_validators<I>(&mut self, validators: I)
-    where
-        I: IntoIterator<Item = TestNode>,
-    {
-        self.validators = validators
-            .into_iter()
-            .enumerate()
-            .map(|(idx, mut node)| {
-                node.change_role(Some(ValidatorId(idx as u16)));
-                node
-            })
-            .collect();
-        self.stored_configuration.validator_keys = self.validators
-            .iter()
-            .cloned()
-            .map(ValidatorKeys::from)
-            .collect();
-        self.update_our_role();
-    }
-
-    /// Returns the configuration for service with the given identifier.
-    pub fn service_config<D>(&self, id: &str) -> D
-    where
-        for<'de> D: Deserialize<'de>,
-    {
-        let value = self.stored_configuration.services.get(id).expect(
-            "Unable to find configuration for service",
-        );
-        serde_json::from_value(value.clone()).unwrap()
-    }
-
-    /// Modifies the configuration of the service with the given identifier.
-    pub fn set_service_config<D>(&mut self, id: &str, config: D)
-    where
-        D: Serialize,
-    {
-        let value = serde_json::to_value(config).unwrap();
-        self.stored_configuration.services.insert(id.into(), value);
-    }
-
-    /// Returns the resulting exonum blockchain configuration.
-    pub fn stored_configuration(&self) -> &StoredConfiguration {
-        &self.stored_configuration
-    }
-
-    fn update_our_role(&mut self) {
-        let validator_id = self.validators
-            .iter()
-            .position(|x| {
-                x.public_keys().service_key == self.us.service_public_key
-            })
-            .map(|x| ValidatorId(x as u16));
-        self.us.validator_id = validator_id;
-    }
-}
-
-/// Kind of public or private REST API of an Exonum node.
-///
-/// `ApiKind` allows to use `get*` and `post*` methods of [`TestKitApi`] more safely.
-///
-/// [`TestKitApi`]: struct.TestKitApi.html
-#[derive(Debug)]
-pub enum ApiKind {
-    /// `api/system` endpoints of the built-in Exonum REST API.
-    System,
-    /// `api/explorer` endpoints of the built-in Exonum REST API.
-    Explorer,
-    /// Endpoints corresponding to a service with the specified string identifier.
-    Service(&'static str),
-}
-
-impl ApiKind {
-    fn into_prefix(self) -> String {
-        match self {
-            ApiKind::System => "api/system".to_string(),
-            ApiKind::Explorer => "api/explorer".to_string(),
-            ApiKind::Service(name) => format!("api/services/{}", name),
-        }
-    }
-}
-
-/// API encapsulation for the testkit. Allows to execute and synchronously retrieve results
-/// for REST-ful endpoints of services.
-pub struct TestKitApi {
-    public_handler: Chain,
-    private_handler: Chain,
-    api_sender: ApiSender,
-}
-
-impl fmt::Debug for TestKitApi {
-    fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
-        f.debug_struct("TestKitApi").finish()
-    }
-}
-
-impl TestKitApi {
-    /// Creates a new instance of API.
-    fn new(testkit: &TestKit) -> Self {
-        use std::sync::Arc;
-
-        let blockchain = &testkit.blockchain;
-        let api_state = SharedNodeState::new(10_000);
-
-        TestKitApi {
-            public_handler: create_public_api_handler(
-                blockchain.clone(),
-                Arc::clone(&testkit.mempool),
-                api_state.clone(),
-                &testkit.api_config,
-            ),
-
-            private_handler: create_private_api_handler(
-                blockchain.clone(),
-                api_state,
-                testkit.api_sender.clone(),
-            ),
-
-            api_sender: testkit.api_sender.clone(),
-        }
-    }
-
-    /// Returns the mounting point for public APIs. Useful for intricate testing not covered
-    /// by `get*` and `post*` functions.
-    pub fn public_handler(&self) -> &Chain {
-        &self.public_handler
-    }
-
-    /// Returns the mounting point for private APIs. Useful for intricate testing not covered
-    /// by `get*` and `post*` functions.
-    pub fn private_handler(&self) -> &Chain {
-        &self.private_handler
-    }
-
-    /// Sends a transaction to the node via `ApiSender`.
-    pub fn send<T>(&self, transaction: T)
-    where
-        T: Into<Box<Transaction>>,
-    {
-        self.api_sender.send(transaction.into()).expect(
-            "Cannot send transaction",
-        );
-    }
-
-    fn get_internal<H, D>(handler: &H, endpoint: &str, expect_error: bool, is_public: bool) -> D
-    where
-        H: Handler,
-        for<'de> D: Deserialize<'de>,
-    {
-        let status_class = if expect_error {
-            StatusClass::ClientError
-        } else {
-            StatusClass::Success
-        };
-
-        let url = format!("http://localhost:3000/{}", endpoint);
-        let resp = request::get(&url, Headers::new(), handler);
-        let resp = if expect_error {
-            // Support either "normal" or erroneous responses.
-            // For example, `Api.not_found_response()` returns the response as `Ok(..)`.
-            match resp {
-                Ok(resp) => resp,
-                Err(IronError { response, .. }) => response,
-            }
-        } else {
-            resp.expect("Got unexpected `Err(..)` response")
-        };
-
-        if let Some(ref status) = resp.status {
-            if status.class() != status_class {
-                panic!("Unexpected response status: {:?}", status);
-            }
-        } else {
-            panic!("Response status not set");
-        }
-
-        let resp = response::extract_body_to_string(resp);
-
-        let publicity = if is_public { "" } else { " private" };
-        trace!("GET{} {}\nResponse:\n{}\n", publicity, endpoint, resp);
-
-        serde_json::from_str(&resp).unwrap()
-    }
-
-    /// Gets information from a public endpoint of the node.
-    ///
-    /// # Panics
-    ///
-    /// - Panics if an error occurs during request processing (e.g., the requested endpoint is
-    ///  unknown), or if the response has a non-20x response status.
-    pub fn get<D>(&self, kind: ApiKind, endpoint: &str) -> D
-    where
-        for<'de> D: Deserialize<'de>,
-    {
-        TestKitApi::get_internal(
-            &self.public_handler,
-            &format!("{}/{}", kind.into_prefix(), endpoint),
-            false,
-            true,
-        )
-    }
-
-    /// Gets information from a private endpoint of the node.
-    ///
-    /// # Panics
-    ///
-    /// - Panics if an error occurs during request processing (e.g., the requested endpoint is
-    ///  unknown), or if the response has a non-20x response status.
-    pub fn get_private<D>(&self, kind: ApiKind, endpoint: &str) -> D
-    where
-        for<'de> D: Deserialize<'de>,
-    {
-        TestKitApi::get_internal(
-            &self.private_handler,
-            &format!("{}/{}", kind.into_prefix(), endpoint),
-            false,
-            false,
-        )
-    }
-
-    /// Gets an error from a public endpoint of the node.
-    ///
-    /// # Panics
-    ///
-    /// - Panics if the response has a non-40x response status.
-    pub fn get_err<D>(&self, kind: ApiKind, endpoint: &str) -> D
-    where
-        for<'de> D: Deserialize<'de>,
-    {
-        TestKitApi::get_internal(
-            &self.public_handler,
-            &format!("{}/{}", kind.into_prefix(), endpoint),
-            true,
-            true,
-        )
-    }
-
-    fn post_internal<H, T, D>(handler: &H, endpoint: &str, data: &T, is_public: bool) -> D
-    where
-        H: Handler,
-        T: Serialize,
-        for<'de> D: Deserialize<'de>,
-    {
-        let url = format!("http://localhost:3000/{}", endpoint);
-        let body = serde_json::to_string(&data).expect("Cannot serialize data to JSON");
-        let resp = request::post(
-            &url,
-            {
-                let mut headers = Headers::new();
-                headers.set(ContentType::json());
-                headers
-            },
-            &body,
-            handler,
-        ).expect("Cannot send data");
-
-        let resp = response::extract_body_to_string(resp);
-
-        let publicity = if is_public { "" } else { " private" };
-        trace!(
-            "POST{} {}\nBody: \n{}\nResponse:\n{}\n",
-            publicity,
-            endpoint,
-            body,
-            resp
-        );
-
-        serde_json::from_str(&resp).expect("Cannot parse result")
-    }
-
-    /// Posts a transaction to the service using the public API. The returned value is the result
-    /// of synchronous transaction processing, which includes running the API shim
-    /// and `Transaction.verify()`. `Transaction.execute()` is not run until the transaction
-    /// gets to a block via one of `create_block*()` methods.
-    ///
-    /// # Panics
-    ///
-    /// - Panics if an error occurs during request processing (e.g., the requested endpoint is
-    ///  unknown).
-    pub fn post<T, D>(&self, kind: ApiKind, endpoint: &str, transaction: &T) -> D
-    where
-        T: Serialize,
-        for<'de> D: Deserialize<'de>,
-    {
-        TestKitApi::post_internal(
-            &self.public_handler,
-            &format!("{}/{}", kind.into_prefix(), endpoint),
-            transaction,
-            true,
-        )
-    }
-
-    /// Posts a transaction to the service using the private API. The returned value is the result
-    /// of synchronous transaction processing, which includes running the API shim
-    /// and `Transaction.verify()`. `Transaction.execute()` is not run until the transaction
-    /// gets to a block via one of `create_block*()` methods.
-    ///
-    /// # Panics
-    ///
-    /// - Panics if an error occurs during request processing (e.g., the requested endpoint is
-    ///  unknown).
-    pub fn post_private<T, D>(&self, kind: ApiKind, endpoint: &str, transaction: &T) -> D
-    where
-        T: Serialize,
-        for<'de> D: Deserialize<'de>,
-    {
-        TestKitApi::post_internal(
-            &self.private_handler,
-            &format!("{}/{}", kind.into_prefix(), endpoint),
-            transaction,
-            false,
-        )
-    }
 }
 
 #[test]
